+++ lib/src/utils/base_digit_separated.dart
@@ -21,7 +21,6 @@ abstract class BaseDigitsSeparated extends StatelessWidget {
     required this.countUp,
     required this.slideAnimationDuration,
     required this.separator,
-    required this.fade,
     required this.hideFirstDigitZero,
     required this.showSeparator,
     this.digitTitle,
@@ -48,7 +47,7 @@ abstract class BaseDigitsSeparated extends StatelessWidget {
   final bool countUp;
   final Duration slideAnimationDuration;
   final String separator;
-  final bool fade, hideFirstDigitZero, showSeparator;
+  final bool hideFirstDigitZero, showSeparator;
   final String? digitTitle;
   final EdgeInsets? separatorPadding;
   final EdgeInsets? digitTitlePadding;