+++ lib/src/utils/text_animation.dart
@@ -9,7 +9,6 @@ class TextAnimation extends StatefulWidget {
     required this.textStyle,
     required this.slideDirection,
     required this.slideAnimationDuration,
-    required this.fade,
     this.curve = Curves.easeOut,
     this.countUp = true,
     this.digitsNumber,
@@ -26,7 +25,7 @@ class TextAnimation extends StatefulWidget {
   final bool countUp;
   final Duration slideAnimationDuration;
   final List<String>? digitsNumber;
-  final bool hideFirstDigitZero, isLastDigit, fade;
+  final bool hideFirstDigitZero, isLastDigit;
 
   @override
   _TextAnimationState createState() => _TextAnimationState();
@@ -107,7 +106,6 @@ class _TextAnimationState extends State<TextAnimation> with SingleTickerProvider
   @override
   Widget build(BuildContext context) {
     final isUp = widget.slideDirection == SlideDirection.up;
-    final clipBehavior = widget.fade ? Clip.none : Clip.hardEdge;
     return AnimatedBuilder(
       animation: _animationController,
       builder: (context, _) {
@@ -127,7 +125,6 @@ class _TextAnimationState extends State<TextAnimation> with SingleTickerProvider
                     slideDirection: widget.slideDirection,
                     percentage: _offsetAnimationOne.value.dy,
                   ),
-                  clipBehavior: clipBehavior,
                   child: Text(
                     digit(nextValue),
                     style: widget.textStyle,
@@ -143,7 +140,6 @@ class _TextAnimationState extends State<TextAnimation> with SingleTickerProvider
                     slideDirection: widget.slideDirection,
                     percentage: _offsetAnimationTwo.value.dy,
                   ),
-                  clipBehavior: clipBehavior,
                   child: Text(
                     digit(currentValue),
                     style: widget.textStyle,