+++ CHANGELOG.md
@@ -1,17 +1,22 @@
-## 0.4.2
+## 0.5.2
 * Added text overflow helper
 
-## 0.4.1
+## 0.5.1
 * Added backdrop filter option
 * Added hide first digit zero option
 * Added option to show digit title underneath the \(normal\) digit
 * Code clean-up
 
-## 0.4.0
+## 0.5.0
 * Added option to show digit title underneath the separated digit
 * Added option to hide separator
 * Upgraded package versions
 * Code refactor and clean-up
+
+## 0.4.0
+* Remove deprecated API
+* `duration` isn't longer required when override duration with `stream_duration`
+* bump `stream_duration` package to `0.3.0`
 
 ## 0.3.4
 * Readme add example control duration